requirements/production.txt
Django==5.1.6
mysqlclient==2.2.4
django-crispy-forms==2.3
crispy-bootstrap5==2024.10
django-jazzmin==3.0.1
boto3==1.37.5
django-storages==1.14.5
python-decouple==3.8
django-import-export==4.2.0
openpyxl==3.1.5